client: Linux fixes
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@696 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ src/client/src/org/xtreemfs/client/volume.cpp
@@ -33,9 +33,9 @@ Volume::Volume( const YIELD::SocketAddress& dir_sockaddr, const std::string& nam
   : name( name ), flags( flags ), log( log )
 {
   stage_group = new YIELD::SEDAStageGroup( name.c_str() );
-  dir_proxy = DIRProxy::create( stage_group, dir_sockaddr, log, 5 * NS_IN_S, YIELD::Client::RECONNECT_TRIES_MAX_DEFAULT, socket_factory );
+  dir_proxy = DIRProxy::create( stage_group, dir_sockaddr, log, static_cast<uint64_t>( 5 * NS_IN_S ), YIELD::Client::RECONNECT_TRIES_MAX_DEFAULT, socket_factory );
   YIELD::auto_Object<YIELD::URI> mrc_uri = dir_proxy->getVolumeURIFromVolumeName( name );
-  mrc_proxy = MRCProxy::create( stage_group, *mrc_uri, log, 5 * NS_IN_S, YIELD::Client::RECONNECT_TRIES_MAX_DEFAULT, socket_factory );
+  mrc_proxy = MRCProxy::create( stage_group, *mrc_uri, log, static_cast<uint64_t>( 5 * NS_IN_S ), YIELD::Client::RECONNECT_TRIES_MAX_DEFAULT, socket_factory );
 }
 
 Volume::~Volume()
